move date format helper to lib/datetime folder
