+++ client/src/components/Notes/Notes.tsx
@@ -1,4 +1,4 @@
-import { formatDate } from "../../lib/format-date";
+import { formatDate } from "../../lib/datetime/format-date";
 import type { TagsData } from "../../types/data.types";
 import type { NoteWithIds } from "../../types/server/note.types";
 import * as S from "./Notes.style";
